pyproject.toml
[project]
name = "voice-diary-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "apscheduler>=3.11.2",
    "httpx>=0.28.1",
    "notion-client>=3.0.0",
    "openai>=2.29.0",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
]
